extract dutystationSearchBox from DutyStation, works but isvalid on wizard page not
